refactor: jwtTokenProvider 의존성 역전
